+++ CHANGELOG.md
@@ -7,3 +7,4 @@
 - Added importWallet() to import account using privateKey
 - Added initial test
 - Added functionality to sign message
+- Added get balance method to fetch SOL balance of an account